+++ src/main/java/org/hilingual/domain/token/TokenRepository.java
@@ -1,8 +1,8 @@
-package org.hilingual.domain.security;
+package org.hilingual.domain.token;
 
 import org.springframework.data.redis.repository.configuration.EnableRedisRepositories;
 import org.springframework.data.repository.CrudRepository;
 
 @EnableRedisRepositories(basePackages = "org.hilingual.domain.security")
-public interface RefreshTokenRepository extends CrudRepository<RefreshToken, Long> {
+public interface TokenRepository extends CrudRepository<Token, Long> {
 }
